+++ test/libsolidity/semanticTests/errors/require_error_function_pointer_parameter.sol
@@ -1,0 +1,20 @@
+error CustomError(function(uint256) external pure returns (uint256));
+
+contract C
+{
+    function e(uint256 x) external pure returns (uint256)
+    {
+        return x;
+    }
+
+    function f() external view
+    {
+        // more than one stack slot
+        require(false, CustomError(this.e));
+    }
+}
+
+// ====
+// compileViaYul: true
+// ----
+// f() -> FAILURE, hex"271b1dfa", hex"c06afe3a8444fc0004668591e8306bfb9968e79ef37cdc8e0000000000000000"