+++ test/libsolidity/syntaxTests/errors/selector_on_instance.sol
@@ -4,4 +4,4 @@ contract C {
     bytes4 t = E().selector;
 }
 // ----
-// TypeError 9582: (40-52): Member "selector" not found or not visible after argument-dependent lookup in tuple().
+// TypeError 9582: (40-52): Member "selector" not found or not visible after argument-dependent lookup in error.