+++ Changelog.md
@@ -1,6 +1,7 @@
 ### 0.8.26 (unreleased)
 
 Language Features:
+ * Introduce a new overload ``require(bool, Error)`` that allows usage of ``require`` functions with custom errors. This feature is available in the ``via-ir`` pipeline only.
 
 
 Compiler Features: